правки по ДЗ
