+++ index.html
@@ -234,7 +234,7 @@
       </section>
     </template>
 
-<script type="module" src="main.js"></script>
+<script type="module" src="../js/main.js"></script>
 
 </body>
 </html>
