+++ js/util.js
@@ -1,42 +1,3 @@
-/* eslint-disable no-console */
-// вычисляем случайный ID
-const id = Array.from({ length: 25 },(_, i) => i + 1);
-// вычисляем случайный url
-// eslint-disable-next-line no-template-curly-in-string
-const url = Array.from({ length: 25 }, (_, i) => ({ url: 'photos/${ i + 1 }.jpg' }));
-// Случайное описание
-const descriptions = [];
-for (let i = 1; i <= 25; i++) {
-  // eslint-disable-next-line no-template-curly-in-string
-  descriptions.push({ description: 'Описание фотографии №${i}' });
-}
-const likes = [];
-for (let i = 0; i < 25; i++) {
-  likes.push({ likes: Math.floor(Math.random() * (200 - 15 + 1)) + 15 });
-}
-const messages = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-// Массив случайных имён
-const names = [
-  'Алексей',
-  'Мария',
-  'Дмитрий',
-  'Елена',
-  'Сергей',
-  'Анна',
-  'Иван',
-  'Ольга',
-  'Николай',
-  'Татьяна'
-];
-
 // Функция: случайное целое число в диапазоне [min, max]
 function getRandomInt(min, max) {
   return Math.floor(Math.random() * (max - min + 1)) + min;
@@ -88,5 +49,5 @@ function generateComments() {
 comments = generateComments();
 //console.log(comments);
 
-export {id, url, descriptions, likes, comments};
+export {getRandomInt,generateComments}
